Fixed memory corruption related to problem objects.
On the C side svm_model keeps a pointer to svm_problem. As documented
this means that the lisp wrapper for the model must keep a reference
to the wrapper of the problem since deallocation of wrapped objects is
tied to the life of their wrappers. The mechanism to do this was there
but it was broken.